scripts/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "grpcio>=1.72.1",
    "orjson>=3.10.18",
    "pygit2>=1.18.0",
    "swh-graph>=6.7.6",
]
